fixed another type issue with mips

+++ mercury/include/types.h
@@ -47,7 +47,7 @@ typedef uint32_t u32_t;
 typedef uint16_t u16_t;
 typedef uint8_t  u8_t;
 
-typedef int64_t s64_t;
+typedef long long int s64_t;
 typedef int32_t s32_t;
 typedef int16_t s16_t;
 typedef int8_t  s8_t;
